+++ Src/Alco/Serialization/Binary/ReferenceContext.cs
@@ -9,15 +9,15 @@ public class ReferenceContext
 {
     public const string SerializeKey = "$id";
 
-    protected uint _nextId = 1;
-    private readonly ConcurrentDictionary<object, uint> _objectToId = new();
-    private readonly ConcurrentDictionary<uint, object> _idToObject = new();
+    protected ulong _nextId = 1;
+    private readonly ConcurrentDictionary<object, ulong> _objectToId = new();
+    private readonly ConcurrentDictionary<ulong, object> _idToObject = new();
 
     /// <summary>
     /// Gets or assigns a unique ID for the specified object within this context.
     /// Override to customize ID assignment (e.g. cross-context references).
     /// </summary>
-    public virtual uint GetId(object obj)
+    public virtual ulong GetId(object obj)
     {
         return _objectToId.GetOrAdd(obj, AddReference);
     }
@@ -27,7 +27,7 @@ public class ReferenceContext
     /// Override to route registration through a shared registry (e.g. cross-context),
     /// keeping the object→id table consistent with <see cref="GetId"/>'s allocation source.
     /// </summary>
-    public virtual void SetReference(uint id, object obj)
+    public virtual void SetReference(ulong id, object obj)
     {
         if (id == 0)
             return;
@@ -40,12 +40,12 @@ public class ReferenceContext
     /// Attempts to resolve an object by its ID.
     /// Override to add fallback resolution (e.g. cross-context registry).
     /// </summary>
-    public virtual bool TryGetReference(uint id, [NotNullWhen(true)] out object? obj)
+    public virtual bool TryGetReference(ulong id, [NotNullWhen(true)] out object? obj)
     {
         return _idToObject.TryGetValue(id, out obj);
     }
 
-    private uint AddReference(object obj)
+    private ulong AddReference(object obj)
     {
         return Interlocked.Increment(ref _nextId);
     }
